Adjusted Code from ESD to have no conflicts.
