+++ web/src/components/Front.jsx
@@ -4,8 +4,8 @@ import { Text } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { ROUTES } from "../constants";
 import ItemCards from "./ItemCards";
+import ItemDescript from "./ItemDescript";
 import Navbar from "./Navbar";
-import ItemDescription from "../pages/ItemDescription";
 
 const Front = () => {
   const Stack = createStackNavigator();
@@ -23,7 +23,7 @@ const Front = () => {
         />
         <Stack.Screen
           name={ROUTES.ITEM_DESCRIPTION}
-          component={ItemDescription}
+          component={ItemDescript}
           options={{ headerShown: false }}
         />
       </Stack.Navigator>
